moved stopArea list to stopAreaView
